backend/requirements.txt
fastapi
uvicorn
watchfiles
psycopg[binary]
python-dotenv
pydantic==2.4.0
pydantic-settings==2.0.3
bcrypt
python-jose[cryptography]
email-validator
sqlalchemy[asyncio]
asyncpg
alembic
python-multipart
psutil
apscheduler
redis
openpyxl
